refactor(markdown): extract CopyButton and its icons from CodeBlockComponent
CodeBlockComponent housed a second component (CopyButton) plus two inline
SVGs, against the one-concern-per-file rule. Move CopyButton to its own
file and the check/copy glyphs to icons/CheckIcon and icons/CopyIcon,
matching the icon convention. CodeBlockComponent now only routes fenced
blocks to the right renderer.
